README: Replace screenshots of tables with Markdown table
The screenshots are from tables with text, which are not easily searchable,
are bigger in size than needed – increasing load times – and are in a
resolution, causing them to be blurry on HiDPI displays.  Therefore, use a
Markdown table instead solving all the issues above, and delete the images
from the repository.

The SLES service pack version differs in the two screenshots: SP2 vs SP3.
Go for *SP3*.

Resolves: https://github.com/RadeonOpenCompute/ROCm/issues/1591